backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
pillow==10.1.0
Pylance==0.8.14
numpy==1.26.1
pyarrow==14.0.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
django-filter==23.3